+++ plone/formwidget/geolocation/interfaces.py
@@ -1,8 +1,11 @@
-from plone.formwidget.geolocation import _
 from z3c.form.interfaces import IWidget
 from zope import schema
+from zope.i18nmessageid import MessageFactory
 from zope.interface import Interface
 from zope.schema.interfaces import IObject
+
+
+_ = MessageFactory('plone.formwidget.geolocation')
 
 
 class IGeolocation(Interface):